+++ docs/reference/index.rst
@@ -8,4 +8,5 @@ Reference
 
     linedraw
     brachiograph
+    turtle_draw
     community-resources
