+++ docs/reference/turtle_draw.rst
@@ -1,0 +1,10 @@
+``turtle_draw.py``
+==========================
+
+..  module:: turtle_draw
+
+..  autoclass:: PantoGraphTurtle
+    :members:
+
+..  autoclass:: BrachioGraphTurtle
+    :members:
