+++ turtle_draw.py
@@ -4,58 +4,7 @@ from turtle import *
 import math
 
 
-class BrachioGraphTurtle(Turtle):
-
-    def __init__(self,
-        inner_arm=8,          # the length of the inner arm (blue)
-        shoulder_centre_angle=0,  # the starting angle of the inner arm, relative to straight ahead
-        shoulder_sweep=180,     # the arc covered by the shoulder motor
-
-        outer_arm=8,          # the length of the outer arm (red)
-        elbow_centre_angle=90,  # the centre of the outer arm relative to the inner arm
-        elbow_sweep=180,        # the arc covered by the elbow motor
-
-        window_size=800,        # width and height of the turtle canvas
-        speed=0                 # how fast to draw
-        ):
-
-        self.inner_arm = inner_arm
-        self.outer_arm = outer_arm
-        self.shoulder_centre_angle = shoulder_centre_angle
-        self.shoulder_sweep = shoulder_sweep
-        self.elbow_centre_angle = elbow_centre_angle
-        self.elbow_sweep = elbow_sweep
-        self.window_size = window_size
-
-        # some basic dimensions of the drawing area
-
-        grid_size = self.window_size / 1.05  # the grid is a little smaller than the window
-
-        # scale the plotter dimensions to fill the screen
-        self.multiplier = grid_size / 2 / (self.inner_arm + self.outer_arm)
-
-        self.reach = self.inner_arm + self.outer_arm  # the maximum possible distance the arms could reach
-        self.draw_reach = self.reach * self.multiplier * 1.05  # maximum possible drawing reacg
-
-        # set up the screen for the turtle
-
-        self.screen = Screen()
-        self.screen.mode("logo")
-        self.screen.title(f"inner length {self.inner_arm}cm • centre {self.shoulder_centre_angle}˚ • sweep {self.shoulder_sweep}˚  •  outer length {self.outer_arm}cm • centre {self.elbow_centre_angle}˚ • sweep {self.elbow_sweep}˚")
-        self.screen.setup(width=window_size, height=window_size)
-
-        super().__init__()
-
-        self.speed(0)
-        self.hideturtle()
-        self.screen.tracer(speed, 0)
-
-
-    def simple_title(self, title=""):
-        title = title or "BrachioGraph, multiple values"
-        self.screen.title(title)
-
-
+class AbstractTurtle:
     # ----------------- grid drawing methods -----------------
 
     def draw_grid(self):
@@ -88,6 +37,117 @@ class BrachioGraphTurtle(Turtle):
                     self.goto(- self.reach * self.multiplier, i * self.multiplier)
                     self.write(i, move=False, font=("Helvetica", 16, "bold"))
 
+
+class PantoGraphTurtle(Turtle, AbstractTurtle):
+    """A turtle-graphics implementation of a PantoGraph. Instantiate your ``PantoGraph`` with
+    ``turtle=True`` to create a turtle version of it, that copies everything the PantoGraph does.
+    """
+
+    def __init__(self,
+        driver=8,
+        follower=8,
+        motor_1_pos = -1.5,
+        motor_2_pos = 1.5,
+        window_size=800,            # width and height of the turtle canvas
+        speed=0,                    # how fast to draw
+
+        motor_1_centre_angle=0,     # the starting angle of the first arm, relative to straight ahead
+        motor_2_centre_angle=90,    # the centre of the second arm relative to the inner arm
+
+        motor_1_sweep=180,          # the arc covered by the first motor
+        motor_2_sweep=180,          # the arc covered by the second motor
+    ):
+
+        # set the pantograph geometry
+        self.driver = driver
+        self.follower = follower
+
+        self.motor_1_pos = motor_1_pos
+        self.motor_2_pos = motor_2_pos
+
+        self.motor_1_centre_angle = motor_1_centre_angle
+        self.motor_2_centre_angle = motor_2_centre_angle
+
+        self.motor_1_sweep = motor_1_sweep
+        self.motor_2_sweep = motor_2_sweep
+
+        # some basic dimensions of the drawing area
+
+        self.window_size = window_size
+        grid_size = self.window_size / 1.05  # the grid is a little smaller than the window
+
+        # scale the plotter dimensions to fill the screen
+        self.reach = self.driver + self.follower  # the maximum possible distance the arms could reach
+        self.multiplier = grid_size / 2 / self.reach
+        self.draw_reach = self.reach * self.multiplier * 1.05  # maximum possible drawing reacg
+
+        # set up the screen for the turtle
+
+        self.screen = Screen()
+        self.screen.mode("logo")
+        self.screen.title(f"driver length {self.driver}cm • centre {self.motor_1_centre_angle}˚ • sweep {self.motor_1_sweep}˚  •  follower length {self.follower}cm • centre {self.motor_2_centre_angle}˚ • sweep {self.motor_2_sweep}˚")
+        self.screen.setup(width=window_size, height=window_size)
+
+        super().__init__()
+
+        self.speed(0)
+        self.screen.tracer(speed, 0)
+
+
+class BrachioGraphTurtle(Turtle, AbstractTurtle):
+    """A turtle-graphics implementation of a BrachioGraph. Instantiate your ``BrachioGraph`` with
+    ``turtle=True`` to create a turtle version of it, that copies everything the BrachioGraph does.
+    """
+
+    def __init__(self,
+        inner_arm=8,
+        outer_arm=8,
+        window_size=800,
+        speed=0,
+
+        shoulder_centre_angle=0,    # the starting angle of the inner arm, relative to straight ahead
+        elbow_centre_angle=90,      # the centre of the outer arm relative to the inner arm
+
+        shoulder_sweep=180,         # the arc covered by the shoulder motor
+        elbow_sweep=180,            # the arc covered by the elbow motor
+        ):
+
+        self.inner_arm = inner_arm
+        self.outer_arm = outer_arm
+
+        self.shoulder_centre_angle = shoulder_centre_angle
+        self.shoulder_sweep = shoulder_sweep
+
+        self.elbow_centre_angle = elbow_centre_angle
+        self.elbow_sweep = elbow_sweep
+
+        self.window_size = window_size
+
+        # some basic dimensions of the drawing area
+
+        grid_size = self.window_size / 1.05  # the grid is a little smaller than the window
+
+        # scale the plotter dimensions to fill the screen
+        self.reach = self.inner_arm + self.outer_arm  # the maximum possible distance the arms could reach
+        self.multiplier = grid_size / 2 / self.reach
+        self.draw_reach = self.reach * self.multiplier * 1.05  # maximum possible drawing reacg
+
+        # set up the screen for the turtle
+
+        self.screen = Screen()
+        self.screen.mode("logo")
+        self.screen.title(f"inner length {self.inner_arm}cm • centre {self.shoulder_centre_angle}˚ • sweep {self.shoulder_sweep}˚  •  outer length {self.outer_arm}cm • centre {self.elbow_centre_angle}˚ • sweep {self.elbow_sweep}˚")
+        self.screen.setup(width=window_size, height=window_size)
+
+        super().__init__()
+
+        self.speed(0)
+        self.screen.tracer(speed, 0)
+
+
+    def simple_title(self, title=""):
+        title = title or "BrachioGraph, multiple values"
+        self.screen.title(title)
 
 
     # ----------------- arc drawing methods -----------------
